REF: Rename remaining private modules to _lowercase.py convention
