+++ MANIFEST.in
@@ -1,2 +1,2 @@
-recursive-include skordinal/classifiers/libsvmRank *.h *.hpp
-recursive-include skordinal/classifiers/svorex *.h
+recursive-include skordinal/classifiers/src/libsvmrank *.h *.hpp
+recursive-include skordinal/classifiers/src/svorex *.h
