+++ skordinal/datasets/__init__.py
@@ -13,12 +13,12 @@ from ._loaders import (
 )
 
 __all__ = [
-    "get_data_path",
-    "dataset_exists",
-    "is_undivided",
-    "has_unseeded_split",
-    "has_seeded_split",
     "check_ambiguity",
+    "dataset_exists",
+    "get_data_path",
+    "has_seeded_split",
+    "has_unseeded_split",
+    "is_undivided",
     "load_datafile",
     "load_dataset",
     "shuffle_data",
