+++ skordinal/preprocessing/__init__.py
@@ -1,5 +1,5 @@
 """Preprocessing module."""
 
-from .scalers import apply_scaling, minmax_scale, standardize
+from ._scalers import apply_scaling, minmax_scale, standardize
 
 __all__ = ["apply_scaling", "minmax_scale", "standardize"]
