+++ skordinal/results/_results.py
