+++ skordinal/testing/__init__.py
@@ -1,6 +1,6 @@
 """Testing utilities module."""
 
-from .testing_utils import (
+from ._testing_utils import (
     TEST_DATASETS_DIR,
     TEST_PREDICTIONS_DIR,
     TEST_PROJECT_ROOT,
@@ -8,8 +8,8 @@ from .testing_utils import (
 )
 
 __all__ = [
-    "TEST_PROJECT_ROOT",
     "TEST_DATASETS_DIR",
     "TEST_PREDICTIONS_DIR",
+    "TEST_PROJECT_ROOT",
     "TEST_RANDOM_STATE",
 ]
